make it easier to add more elsewhere platforms
