+++ liberapay/elsewhere/bitbucket.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from aspen import Response
-from liberapay.elsewhere import PlatformOAuth1
+from liberapay.elsewhere._base import PlatformOAuth1
 from liberapay.elsewhere._extractors import any_key, key, not_available
 from liberapay.elsewhere._paginators import keys_paginator
