+++ liberapay/elsewhere/google.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from liberapay.elsewhere import PlatformOAuth2
+from liberapay.elsewhere._base import PlatformOAuth2
 from liberapay.elsewhere._extractors import any_key, key
 from liberapay.elsewhere._paginators import query_param_paginator
