+++ requirements_tests.txt
@@ -8,7 +8,12 @@ pytest-cov==1.8.1
 
 execnet==1.3.0
 pytest-cache==1.0
-pyflakes==0.8.1
+
+mccabe==0.3.1
+pep8==1.7.0
+pyflakes==1.0.0
+flake8==2.5.0
+
 mock==1.0.1
 
 PyYAML==3.11
