+++ sql/app-conf-defaults.sql
@@ -1,27 +1,28 @@
 INSERT INTO app_conf (key, value) VALUES
     ('bitbucket_callback', '"http://127.0.0.1:8339/on/bitbucket/associate"'::jsonb),
-    ('bitbucket_consumer_key', '"Fu9aQHB64WdVR7UxY4"'::jsonb),
-    ('bitbucket_consumer_secret', '"Y5G6A9BaWDxn2gLKZwwkrGtVE3Zjd7y7"'::jsonb),
-    ('bountysource_api_host', '"https://staging-api.bountysource.com"'::jsonb),
-    ('bountysource_api_secret', '""'::jsonb),
+    ('bitbucket_id', '"Fu9aQHB64WdVR7UxY4"'::jsonb),
+    ('bitbucket_secret', '"Y5G6A9BaWDxn2gLKZwwkrGtVE3Zjd7y7"'::jsonb),
+    ('bountysource_api_url', '"https://staging-api.bountysource.com"'::jsonb),
+    ('bountysource_auth_url', '"https://staging.bountysource.com"'::jsonb),
     ('bountysource_callback', '"http://127.0.0.1:8339/on/bountysource/associate"'::jsonb),
-    ('bountysource_www_host', '"https://staging.bountysource.com"'::jsonb),
+    ('bountysource_id', 'null'::jsonb),
+    ('bountysource_secret', '""'::jsonb),
     ('cache_static', 'false'::jsonb),
     ('check_db_every', '600'::jsonb),
     ('compress_assets', 'false'::jsonb),
     ('dequeue_emails_every', '60'::jsonb),
-    ('facebook_app_id', '"1418954898427187"'::jsonb),
-    ('facebook_app_secret', '"3bcb5dc6ce821e5202870c1e6ef5bbc4"'::jsonb),
     ('facebook_callback', '"http://localhost:8339/on/facebook/associate"'::jsonb),
+    ('facebook_id', '"1418954898427187"'::jsonb),
+    ('facebook_secret', '"3bcb5dc6ce821e5202870c1e6ef5bbc4"'::jsonb),
     ('github_callback', '"http://127.0.0.1:8339/on/github/associate"'::jsonb),
-    ('github_client_id', '"18891d01e40e5aef93b8"'::jsonb),
-    ('github_client_secret', '"46f75669895e96029d57b64832d6f2c8e6291a0e"'::jsonb),
+    ('github_id', '"18891d01e40e5aef93b8"'::jsonb),
+    ('github_secret', '"46f75669895e96029d57b64832d6f2c8e6291a0e"'::jsonb),
     ('gitlab_callback', '"http://127.0.0.1:8339/on/gitlab/associate"'::jsonb),
     ('gitlab_id', '"c692d729f207df233487857f2909fa06a3232f9ae704e3889b594c7e042cc132"'::jsonb),
     ('gitlab_secret', '"e9b6a5300f3e2a23bee529c4b3041889d2d705d4f5aa6265b1f398757c7b7539"'::jsonb),
     ('google_callback', '"http://localhost:8339/on/google/associate"'::jsonb),
-    ('google_client_id', '"547044792904-jikvtdgeigpot9ek7tsb5660t8rp9p77.apps.googleusercontent.com"'::jsonb),
-    ('google_client_secret', '"7AkgPekyWr6stQWdM6y1TtV6"'::jsonb),
+    ('google_id', '"547044792904-jikvtdgeigpot9ek7tsb5660t8rp9p77.apps.googleusercontent.com"'::jsonb),
+    ('google_secret', '"7AkgPekyWr6stQWdM6y1TtV6"'::jsonb),
     ('linuxfr_callback', '"https://redirect-to-http.oy.lc/127.0.0.1:8339/on/linuxfr/associate"'::jsonb),
     ('linuxfr_id', '"c574b5f45ce054a0750a3764b3ff8ff2c9f88fe36611272a0bf5e4e07bd3c0bf"'::jsonb),
     ('linuxfr_secret', '"8c518595790487015cd0c33be2d04939946f99aad33c86a6af20a99bc749fb3b"'::jsonb),
@@ -33,10 +34,10 @@ INSERT INTO app_conf (key, value) VALUES
     ('openstreetmap_api_url', '"http://www.openstreetmap.org/api/0.6"'::jsonb),
     ('openstreetmap_auth_url', '"http://www.openstreetmap.org"'::jsonb),
     ('openstreetmap_callback', '"http://127.0.0.1:8339/on/openstreetmap/associate"'::jsonb),
-    ('openstreetmap_consumer_key', '"w4eQbkobmMzpkJFwS4tM16a3lq9AFbcoNCKNcGBE"'::jsonb),
-    ('openstreetmap_consumer_secret', '"W08UgEhxQnh7nMzB7GfSFcqcwPnZMmKMNyxWdcw4"'::jsonb),
+    ('openstreetmap_id', '"w4eQbkobmMzpkJFwS4tM16a3lq9AFbcoNCKNcGBE"'::jsonb),
+    ('openstreetmap_secret', '"W08UgEhxQnh7nMzB7GfSFcqcwPnZMmKMNyxWdcw4"'::jsonb),
     ('password_rounds', '1'::jsonb),
     ('twitter_callback', '"http://127.0.0.1:8339/on/twitter/associate"'::jsonb),
-    ('twitter_consumer_key', '"h8bBZtoPNz63S5RkZdbo9R5zb"'::jsonb),
-    ('twitter_consumer_secret', '"Jye64vkWxa2dQu64feTnk0BM3j4JO8ZlTa4EQvMDwrweLkwPaw"'::jsonb),
+    ('twitter_id', '"h8bBZtoPNz63S5RkZdbo9R5zb"'::jsonb),
+    ('twitter_secret', '"Jye64vkWxa2dQu64feTnk0BM3j4JO8ZlTa4EQvMDwrweLkwPaw"'::jsonb),
     ('update_global_stats_every', '300'::jsonb);
